DriftGuard/requirements.txt
# Core Django - Latest stable versions (based on Context7 MCP)
Django>=5.1.0,<6.0
djangorestframework>=3.15.0,<4.0
djangorestframework-simplejwt>=5.3.0,<6.0
django-filter>=25.0,<26.0

# Database
psycopg2-binary>=2.9.7,<3.0
dj-database-url>=2.2.0,<3.0

# Task Queue - Modern versions
celery[redis]>=5.4.0,<6.0
django-celery-beat>=2.6.0,<3.0
django-celery-results>=2.6.0,<3.0

# Machine Learning - Latest versions with pandas 2.x support
scikit-learn>=1.5.0,<2.0
xgboost>=2.1.0,<3.0
pandas>=2.2.0,<3.0
numpy>=2.0.0,<3.0
joblib>=1.4.0,<2.0
imbalanced-learn>=0.12.0,<1.0

# HTTP Client - Modern async support
requests>=2.32.0,<3.0
httpx>=0.27.0,<1.0

# Google Gemini AI - Latest
google-generativeai>=0.8.0,<1.0

# Security & Encryption
cryptography>=43.0.0,<44.0
bcrypt>=4.2.0,<5.0

# Web Server
gunicorn>=23.0.0,<24.0

# Cloud Provider SDKs - Latest versions
boto3>=1.35.0,<2.0
google-cloud-storage>=2.18.0,<3.0
azure-identity>=1.18.0,<2.0
azure-storage-blob>=12.23.0,<13.0

# Monitoring & Observability
sentry-sdk>=2.18.0,<3.0
django-prometheus>=2.4.0,<3.0
structlog>=24.4.0,<25.0

# Development & Testing
pytest>=8.3.0,<9.0
pytest-django>=4.9.0,<5.0
pytest-cov>=5.0.0,<6.0
factory-boy>=3.3.0,<4.0
faker>=27.0.0,<28.0

# Caching & Utilities
redis>=5.2.0,<6.0
python-decouple>=3.8,<4.0
django-cors-headers>=4.6.0,<5.0
django-redis>=5.4.0,<6.0

# Frontend (for dashboard)
django-bootstrap5>=24.1,<25.0

# Documentation
sphinx>=8.0.0,<9.0
sphinx-rtd-theme>=3.0.0,<4.0
